updated part of map
