+++ app/build.gradle
@@ -32,7 +32,11 @@ dependencies {
     compile 'com.firebaseui:firebase-ui:1.2.0'
     compile 'com.google.firebase:firebase-auth:10.0.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    compile 'com.firebase:geofire-android:2.1.1'
+    compile 'com.google.maps.android:android-maps-utils:0.4.3'
+    compile 'com.google.android.gms:play-services-maps:10.2.0'
     testCompile 'junit:junit:4.12'
+
 }
