+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
 
     <application
         android:allowBackup="true"
@@ -56,6 +57,11 @@
         <activity android:name=".PrivateMsgActivity" />
         <activity android:name=".MapActivity" />
         <activity android:name=".NewActivity"></activity>
+
+        <meta-data
+            android:name="com.google.android.geo.API_KEY"
+            android:value="AIzaSyCksJ0rHBb9LzVznlQEc9VXBbxEf2y-hik" />
+
         <service
             android:name=".FetchAddressIntentService"
             android:exported="false"/>
